feat: 적금 상품 가입 시 납입 이력 기록 구현

+++ src/main/java/com/example/reframe/dto/enroll/EnrollForm.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class EnrollForm {
 	private Long periodMonths;
 	private Long paymentAmount;
-	private Long transferDate;
+	private Integer transferDate;
 	
 	private String groupName;
 	private String groupType;
